+++ backend/src/HorseRacing.Domain/Entities/Users/JockeyProfile.cs
@@ -2,7 +2,7 @@ namespace HorseRacing.Domain.Entities;
 
 public class JockeyProfile
 {
-    public int JockeyId { get; set; }
+    public long JockeyId { get; set; }
     public int UserId { get; set; }
     public AppUser? User { get; set; }
     public int ExperienceYears { get; set; }
